requirements.txt
# Core dependencies
google-generativeai>=0.3.1
python-dotenv>=1.0.0

# Web frontend options
streamlit>=1.28.0  # For a simple web interface
flask>=2.3.0       # Alternative web framework
jinja2>=3.1.2      # Templates for Flask

# Additional utilities
requests>=2.31.0   # HTTP requests
pydantic>=2.4.0    # Data validation
